hv_saved_state: move MemoryBlockSaveStruct to defs module
This is a worker process structure from onecore/vm/worker/, not a
hypervisor definition. Move it out of hvdef::save_restore into a
dedicated defs module in hv_saved_state, following the same defs/logic
separation pattern used by hvs_file.
