+++ vm/hv1/hv_saved_state/Cargo.toml
@@ -9,6 +9,7 @@ rust-version.workspace = true
 [dependencies]
 hvdef.workspace = true
 hvs_file.workspace = true
+static_assertions.workspace = true
 thiserror.workspace = true
 zerocopy.workspace = true
