+++ vm/hv1/hv_saved_state/src/lib.rs
@@ -33,6 +33,7 @@
 //! vmrs.finish().unwrap();
 //! ```
 
+mod defs;
 mod partition_state;
 mod vmrs_writer;
